Example of where using an import doesn't work
This adds a line to the test that causes it to fail.  It illustrates
that even with this plugin you unfortunately can't use 'import' directly
from the Jenkinsfile; it appears that the 'import' statements there get
processed before any other lines in the script, and thus the classpath
won't yet be set properly.